tokenizer: SentencePiece Unigram in C, with a measured precision bound
Adds src/tokenizer_sentencepiece.{c,h} to CORE_SOURCES and its model-free
self-test to --self-test, plus tests/test_tokenizer_sp.c and a
make tokenizer-parity target that replays the Python oracle.

Parity is 45,197 cases and 2,116,052 ids across the five languages, clean under
UBSan and ASan with zero leaks. The self-test builds a synthetic ModelProto by
hand with cases chosen to discriminate rather than merely pass: a user-defined
piece whose stored score would lose if it were used instead of the length bonus,
an unused piece that would win if it were not skipped, a minimum score that
changes if byte and control pieces are included, and a trainer-spec unk id that
differs from the unknown piece's index.

The note this was written from said to accumulate in float32 to match upstream.
Measured, that is the worse choice. On the deliberately adversarial
hundred-thousand-character case the accumulated path score reaches about -3e5,
where a float32 step is larger than the gap between competing segmentations, so
ties resolve arbitrarily: six ids out of 102,899 diverged for English, each a
two-character span split the other way, plus one German case. The same substring
tokenized in isolation matches exactly, which is what shows it is precision and
not a bug. With a double accumulator four of the five languages are exact over
the whole corpus and German keeps one divergence on that same case.

Rather than widen a tolerance, the harness states the regime: below 64 KiB any
mismatch fails, above it a divergence is reported loudly and does not, because
finite precision cannot resolve those ties and neither implementation is right.
Nothing in the product reaches that size, since PocketTTS chunks text at fifty
tokens, but the bound is written down so nobody later fixes a non-bug or
loosens the gate to make a real failure pass.

Also rewrites the streaming server note around the finding that
synthesize_slots is a closed batch with no admission point, which reorders that
epic: the global mutex is not the first step, giving the driver a next_job sink
is.

+++ src/tokenizer_sentencepiece.c
@@ -937,7 +937,15 @@ static int sp_normalize(const mynah_sp *sp, const char *text, size_t text_length
  * there is nothing here for a reassociating compiler to regroup. */
 static int sp_viterbi(const mynah_sp *sp, const sp_normalized *n, int *ids, size_t ids_capacity,
                       size_t *out_count, char *error, size_t error_capacity) {
-    float *best = NULL;
+    /* The DP accumulator is double, not float, even though upstream's lattice
+     * stores float scores. On a ~100k-character input the accumulated path
+     * score reaches about -3e5, where a float32 ULP (~0.03) is larger than the
+     * gap between competing segmentations, and ties resolve arbitrarily. In
+     * float this implementation diverged from sentencepiece on 6 ids out of
+     * 102,899 for English and on one German case; in double, four of the five
+     * languages are exact over the whole corpus. Double costs 4 bytes per
+     * character in a prefill-only array. See .work/tokenizer-sentencepiece.md. */
+    double *best = NULL;
     size_t *prev_char = NULL;
     int32_t *prev_id = NULL;
     size_t *order = NULL;
@@ -966,7 +974,7 @@ static int sp_viterbi(const mynah_sp *sp, const sp_normalized *n, int *ids, size
         prev_char[c] = 0;
         prev_id[c] = -1;
     }
-    best[0] = 0.0f;
+    best[0] = 0.0;
 
     for (c = 0; c < n->nchar; c++) {
         const size_t start = n->starts[c];
